move ConflictsResolveDialog to conflict package
